refactor: move arquivos da visão do mestre de jogo para uma pasta
