+++ lib/game_master/locations.dart
